+++ app/Http/Controllers/JoueurController.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Facades\DB;
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 use App\Models\Parente;
+use Illuminate\Support\Facades\Auth;
 
 
 class JoueurController extends Controller
@@ -14,7 +15,9 @@ class JoueurController extends Controller
     // public function __construct()
     // {
     //     $this->middleware('permission:manage players', ['only' => ['create', 'store', 'edit', 'update', 'destroy']]);
+    //     $this->middleware(['role:admin'])->only(['create', 'store', 'edit', 'update', 'destroy']);
     // }
+
 
     public function index()
     {
@@ -24,6 +27,10 @@ class JoueurController extends Controller
 
     public function create()
     {
+        if (!Auth::user()->hasRole('admin')) {
+            abort(403, 'Accès interdit');
+        }
+    
         $parentes = Parente::all();
         return view('joueurs.create', compact('parentes'));
     }
@@ -49,12 +56,19 @@ class JoueurController extends Controller
 
     public function edit(Joueur $joueur)
     {
+        if (!Auth::user()->hasRole('admin')) {
+            abort(403, 'Accès interdit');
+        }
         $parentes = Parente::all();
         return view('joueurs.edit', compact('joueur', 'parentes'));
     }
 
     public function update(Request $request, Joueur $joueur)
     {
+        if (!Auth::user()->hasRole('admin')) {
+            abort(403, 'Accès interdit');
+        }
+
         $request->validate([
             'name' => 'required',
             'position' => 'required',
@@ -63,12 +77,17 @@ class JoueurController extends Controller
         ]);
 
         $joueur->update($request->all());
+        
 
         return redirect()->route('joueurs.index')->with('success', 'Joueur updated successfully.');
     }
 
     public function destroy(Joueur $joueur)
     {
+        if (!Auth::user()->hasRole('admin')) {
+            abort(403, 'Accès interdit');
+        }
+        
         $joueur->delete();
 
         return redirect()->route('joueurs.index')->with('success', 'Joueur deleted successfully.');
